+++ assignment-1/npp-sim/queue.h
@@ -4,12 +4,12 @@
 typedef struct queue {
   node *head;
   node *tail;
-  int capacity;
-  int size;
+  long long capacity;
+  long long size;
 } queue;
 
 void queue_init(queue *q);
-void queue_init_capacity(queue *q, int capacity);
+void queue_init_capacity(queue *q, long long capacity);
 int queue_push(queue *q, packet *p);
 packet *queue_pop(queue *q);
 int queue_empty(queue *q);
